feat(UploadImageBox): Transforma caixa de upload de imagens em componente.
